FunCtrlCht.R v2
renaned R/FunCtrlCht_V2.R to R/FunCtrlCht.R  This version fixes issues identified with how runs.signal was being reported in the earlier version. Additionally minor coding changes to update column headers in final data tables for each chart to be more user friendly. Renamed the input parameters in msr_calc() to be consistent with the other control charting functions.
Added MeanShift.qmd to illustrate the issue with the runs.signal flag and document the behavior of the updated FunCtrlCht.R script including the changes to the final data tables for improved readability.